improvement!: remove all parts using old approach
BREAKING CHANGE: this commit removes all instances where old way of passing repo into functions is used

- CommitText is removed
- OpenRepo is removed
